cleaned up a little -- moved list loading functions to the respository class.... altered the toasts onFlip's so they're a little... prettier?
i think i branch this and just crunch out some of it to do the hearthstone app ---

wow that's tempting.

no no no, we'll get way more out of doing that raw
