+++ app/src/main/java/com/saucefan/stuff/sprint_oop_mdf/view/ItemDetailActivity.kt
@@ -12,7 +12,7 @@ import kotlinx.android.synthetic.main.activity_item_detail.*
 
 class ItemDetailActivity : AppCompatActivity(), ItemDetailFragment.FragmentFavoriteListener {
     override fun flipFavorite(item: AoeTypes) {
-            Toast.makeText(this,"${swapArrayFave(item)}isFavorite: swapArrayFave(item)", Toast.LENGTH_LONG).show()
+            Toast.makeText(this,swapArrayFave(item), Toast.LENGTH_LONG).show()
     }
 
     //lateinit var attachedActivity
